Move base attribute class into the original class to make changes clearer
